fix: harden browser extension security and reduce code duplication
- Add sender validation on background message handler to prevent cross-extension message injection (S1)
- Move host_permissions to optional_host_permissions to require explicit user grant (S2)
- Add API path validation to prevent path traversal attacks (S4)
- Remove empty content_scripts from manifest to reduce permission scope (S5/O4)
- Extract duplicate fetchAccounts into shared lib/fetchAccounts.ts (O2)
- Replace dynamic import with static import in AddAccountForm (O1)
- Remove unused content entry from vite build config
